Reimplement adding a wallet when access is requested and no wallet exists
