+++ src/router/index.js
@@ -45,19 +45,6 @@ router.beforeResolve((to, from, next) => {
     if (store.state.main.wallet === null) {
       if (BUILD_TYPE === MEW_CX) {
         next('/');
-        // ExtensionHelpers.getAccounts(item => {
-        //   const hasStoredWallet = Object.keys(item).filter(key => {
-        //     const newObj = {};
-        //     if (isAddress(key)) {
-        //       return (newObj[key] = item[key]);
-        //     }
-        //   });
-        //   if (hasStoredWallet.length > 0) {
-        //     next('/');
-        //   } else {
-        //     next({ name: 'AccessWalletLayout' });
-        //   }
-        // });
       } else {
         store.dispatch('main/setLastPath', to.path);
         next({ name: 'AccessWalletLayout' });
